add a messages.proto file
This file is kind of the data structure in which we can have fields that
we want in the message.

Currently we are using these fields:
 - x co-ordinate of the bot
 - y co-ordinate of the bot
 - velocity of bot in x-direction
 - velocity of bot in y-direction

This can definitely be extended to incorporate more and more fields as
and when we require. But right now we are testing so these are enough.

Signed-off-by: Pranit Bauva <pranit.bauva@gmail.com>